Clean up driver headers and move ModesConstants to accel.h
