+++ driver/FixedMath/FixedUtil.h
@@ -34,8 +34,6 @@
 // Include numeric types
 #include <linux/types.h>
 
-#include <linux/kernel.h>
-
 // If FP_ASSERT is not custom-defined, then use the standard one
 //#ifndef FP_ASSERT
 //#   include <assert.h>
